Create TipCalculator
Create a web page for calculating tips.

There should be a form that allows the user to enter the Meal amount and the tax Percent as a whole number. It should contain radio buttons for tips of 10, 15 and 20 percent as well as one for other. The other will have a text box where they can enter their custom percent as a whole number. There should be a button to submit and after it is clicked the page should display the amount, the tax, the tip and the total.

Apply a CSS style sheet to the page to make it more presentable.