Create CalenderOfEvents
Create a web page with a Calendar control.

Use a two dimensional array with that contains dates and description of events on that date. It should contain at least 5 or 6 different events. At least one of the dates should have more than one event. (The date will need to be entered as a string and converted to a date for comparison to the date on the calendar.) When a date on the calendar is clicked, it should return a list of the events for that date.

Apply some CSS to format the page: Have a heading one (h1)  element  change its color. Change the color of the event text. Format the Calendar (it renders into a table). You can add any other formatting you wish.