Add files via upload
This "Alert" calendar has events stored, with one date with more than one alert. The date will is entered as a string and converted to a date for comparison to the date on the calendar. When a dates on the calendar is clicked, it returns a list of the events for that date.